@@ -4,17 +4,21 @@ import { ConfigService } from './config.service';
 import { HttpClientModule } from '@angular/common/http';
 
 describe('ConfigService', () => {
-  beforeEach(() => TestBed.configureTestingModule({
-    imports: [
-      HttpClientModule,
-    ],
-    providers: [
-      ConfigService
-    ]
-  }));
+  let service: ConfigService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      imports: [
+        HttpClientModule,
+      ],
+      providers: [
+        ConfigService
+      ]
+    });
+    service = TestBed.inject(ConfigService);
+  });
 
   it('should be created', () => {
-    const service: ConfigService = TestBed.get(ConfigService);
     expect(service).toBeTruthy();
   });
 });
