@@ -4,14 +4,18 @@ import { HttpClientModule } from '@angular/common/http';
 import { ItemsService } from './items.service';
 
 describe('ItemsService', () => {
-  beforeEach(() => TestBed.configureTestingModule({
-    imports: [
-      HttpClientModule
-    ],
-  }));
+  let service: ItemsService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({
+      imports: [
+        HttpClientModule
+      ],
+    });
+    service = TestBed.inject(ItemsService);
+  });
 
   it('should be created', () => {
-    const service: ItemsService = TestBed.get(ItemsService);
     expect(service).toBeTruthy();
   });
 });
