@@ -1,6 +1,7 @@
 import { Component, OnInit } from '@angular/core';
 import { ChartOptions, ChartType, ChartDataSets } from 'chart.js';
-import { Label } from 'ng2-charts';
+import { Color, BaseChartDirective, Label } from 'ng2-charts';
+
 
 @Component({
   selector: 'app-chartjs',
