@@ -1,6 +1,6 @@
 export class Smartphone {
-  name: string;
-  model: string;
-  prize: string;
-  year: string;
-}
+    name: string;
+    model: string;
+    prize: string;
+    year: string;
+  }
