@@ -3,13 +3,15 @@ import { CommonModule } from '@angular/common';
 
 import { ItemsComponent } from './items.component';
 import { ItemsRoutingModule } from './items-routing.module';
+// import { HttpClientModule } from '@angular/common/http';
 import { ItemsService } from './items.service';
 
 @NgModule({
   imports: [
     CommonModule,
     ItemsRoutingModule,
-],
+//    HttpClientModule
+  ],
   exports: [
     ItemsComponent
   ],
