@@ -3,10 +3,14 @@ import { TestBed } from '@angular/core/testing';
 import { SongService } from './song.service';
 
 describe('SongService', () => {
-  beforeEach(() => TestBed.configureTestingModule({}));
+  let service: SongService;
+
+  beforeEach(() => {
+    TestBed.configureTestingModule({});
+    service = TestBed.inject(SongService);
+  });
 
   it('should be created', () => {
-    const service: SongService = TestBed.get(SongService);
     expect(service).toBeTruthy();
   });
 });
