@@ -3,8 +3,8 @@ export const environment = {
   application:
   {
     name: 'angular-starter',
-    angular: 'Angular 8.2.14',
+    angular: 'Angular 9.0.0',
     bootstrap: 'Bootstrap 4.4.1',
-    fontawesome: 'Font Awesome 5.12.0',
+    fontawesome: 'Font Awesome 5.12.1',
   }
 };
